CanTalkAgain fixed.
Some npcs have repeatable dialogues and some don't. I have implemented that.
The dialogue can be triggered by pressing F and can be continued by pressing G. Later on I want to connect it with the controls to change it so spacebar for example, but then the player movement must be frozen. Then the question also calls -> do we want the game to be paused whenever in dialogue?